Improve TestTargetHUD
